// src/API.ts
export enum TripStatus {
  NEW = "NEW",
  ACCEPTED = "ACCEPTED",
  PICKED_UP = "PICKED_UP",
  COMPLETED = "COMPLETED",
  CANCELED = "CANCELED",
}
// ...
export type Trips = {
  __typename: "Trips",
  id: string,
  trip_status?: TripStatus | null,
  pickup_lat?: string | null,
  pickup_lng?: string | null,
  dropoff_lat?: string | null,
  dropoff_lng?: string | null,
  pickup_location_description?: string | null,
  dropoff_location_description?: string | null,
  trip_cost?: number | null,
  TotalQtyItems?: number | null,
  receipt?: string | null,
  duration?: number | null,
  distance?: number | null,
  store_name?: string | null,
  store_image?: string | null,
  user_name?: string | null,
  user_photo?: string | null,
  mobile_number?: string | null,
  driver_name?: string | null,
  driver_photo?: string | null,
  driver_rating?: number | null,
  driver_mobile_number?: string | null,
  driver_car_model?: string | null,
  driver_car_make?: string | null,
  driver_car_color?: string | null,
  driver_car_plate_number?: string | null,
  sub_total?: number | null,
  redeem_points?: number | null,
  service_fee?: number | null,
  delivery_fee?: number | null,
  delivery_discount?: number | null,
  sum_total?: number | null,
  Driver?: Driver | null,
  userID?: string | null,
  storeID: string,
  TripCarts?: ModelTripCartConnection | null,
  Ratings?: ModelRatingConnection | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  tripsDriverId?: string | null,
};

export type Driver = {
  __typename: "Driver",
  id: string,
  name: string,
  email: string,
  image?: string | null,
  phone_number?: string | null,
  rating?: number | null,
  inviteCode?: string | null,
  lng?: string | null,
  lat?: string | null,
  verificationStatus?: DriverRegistrationStatus | null,
  onlineStatus?: boolean | null,
  Cars?: ModelCarConnection | null,
  Ratings?: ModelRatingConnection | null,
  DriverWallets?: ModelDriverWalletConnection | null,
  LicenseDocuments?: ModelLicenseDocumentConnection | null,
  RegistrationDocuments?: ModelRegistrationDocumentConnection | null,
  InsuranceDocuments?: ModelInsuranceDocumentConnection | null,
  fcmToken?: string | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export enum DriverRegistrationStatus {
  NEW = "NEW",
  COMPLETE = "COMPLETE",
  REJECTED = "REJECTED",
}


export type ModelCarConnection = {
  __typename: "ModelCarConnection",
  items:  Array<Car | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelRatingConnection = {
  __typename: "ModelRatingConnection",
  items:  Array<Rating | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};
// ...
export type UpdateDriverMutationVariables = {
  input: UpdateDriverInput,
  condition?: ModelDriverConditionInput | null,
};
// ...
export type RatingsByDriverIDQueryVariables = {
  driverID: string,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelRatingFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type RatingsByDriverIDQuery = {
  ratingsByDriverID?:  {
    __typename: "ModelRatingConnection",
    items:  Array< {
      __typename: "Rating",
      id: string,
      rating?: number | null,
      comment?: string | null,
      driver_image?: string | null,
      driver_name?: string | null,
      customer_name?: string | null,
      customer_image?: string | null,
      tripsID: string,
      driverID: string,
      userID: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};
// ...
export type GetTripsQueryVariables = {
  id: string,
};

export type GetTripsQuery = {
  getTrips?:  {
    __typename: "Trips",
    id: string,
    trip_status?: TripStatus | null,
    pickup_lat?: string | null,
    pickup_lng?: string | null,
    dropoff_lat?: string | null,
    dropoff_lng?: string | null,
    pickup_location_description?: string | null,
    dropoff_location_description?: string | null,
    trip_cost?: number | null,
    TotalQtyItems?: number | null,
    receipt?: string | null,
    duration?: number | null,
    distance?: number | null,
    store_name?: string | null,
    store_image?: string | null,
    user_name?: string | null,
    user_photo?: string | null,
    mobile_number?: string | null,
    driver_name?: string | null,
    driver_photo?: string | null,
    driver_rating?: number | null,
    driver_mobile_number?: string | null,
    driver_car_model?: string | null,
    driver_car_make?: string | null,
    driver_car_color?: string | null,
    driver_car_plate_number?: string | null,
    sub_total?: number | null,
    redeem_points?: number | null,
    service_fee?: number | null,
    delivery_fee?: number | null,
    delivery_discount?: number | null,
    sum_total?: number | null,
    Driver?:  {
      __typename: "Driver",
      id: string,
      name: string,
      email: string,
      image?: string | null,
      phone_number?: string | null,
      rating?: number | null,
      inviteCode?: string | null,
      lng?: string | null,
      lat?: string | null,
      verificationStatus?: DriverRegistrationStatus | null,
      onlineStatus?: boolean | null,
      Cars?:  {
        __typename: "ModelCarConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      Ratings?:  {
        __typename: "ModelRatingConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      DriverWallets?:  {
        __typename: "ModelDriverWalletConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      LicenseDocuments?:  {
        __typename: "ModelLicenseDocumentConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      RegistrationDocuments?:  {
        __typename: "ModelRegistrationDocumentConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      InsuranceDocuments?:  {
        __typename: "ModelInsuranceDocumentConnection",
        nextToken?: string | null,
        startedAt?: number | null,
      } | null,
      fcmToken?: string | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null,
    userID?: string | null,
    storeID: string,
    TripCarts?:  {
      __typename: "ModelTripCartConnection",
      items:  Array< {
        __typename: "TripCart",
        id: string,
        expiry_date: string,
        name: string,
        image: string,
        quantity: number,
        price: number,
        description?: string | null,
        storeID: string,
        userID: string,
        tripsID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    Ratings?:  {
      __typename: "ModelRatingConnection",
      items:  Array< {
        __typename: "Rating",
        id: string,
        rating?: number | null,
        comment?: string | null,
        driver_image?: string | null,
        driver_name?: string | null,
        customer_name?: string | null,
        customer_image?: string | null,
        tripsID: string,
        driverID: string,
        userID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    tripsDriverId?: string | null,
  } | null,
};
// ...
export type GetDriverQueryVariables = {
  id: string,
};

export type GetDriverQuery = {
  getDriver?:  {
    __typename: "Driver",
    id: string,
    name: string,
    email: string,
    image?: string | null,
    phone_number?: string | null,
    rating?: number | null,
    inviteCode?: string | null,
    lng?: string | null,
    lat?: string | null,
    verificationStatus?: DriverRegistrationStatus | null,
    onlineStatus?: boolean | null,
    Cars?:  {
      __typename: "ModelCarConnection",
      items:  Array< {
        __typename: "Car",
        id: string,
        model: string,
        make: string,
        year: number,
        plate_number: string,
        color: string,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    Ratings?:  {
      __typename: "ModelRatingConnection",
      items:  Array< {
        __typename: "Rating",
        id: string,
        rating?: number | null,
        comment?: string | null,
        driver_image?: string | null,
        driver_name?: string | null,
        customer_name?: string | null,
        customer_image?: string | null,
        tripsID: string,
        driverID: string,
        userID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    DriverWallets?:  {
      __typename: "ModelDriverWalletConnection",
      items:  Array< {
        __typename: "DriverWallet",
        id: string,
        amount: number,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    LicenseDocuments?:  {
      __typename: "ModelLicenseDocumentConnection",
      items:  Array< {
        __typename: "LicenseDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    RegistrationDocuments?:  {
      __typename: "ModelRegistrationDocumentConnection",
      items:  Array< {
        __typename: "RegistrationDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    InsuranceDocuments?:  {
      __typename: "ModelInsuranceDocumentConnection",
      items:  Array< {
        __typename: "InsuranceDocument",
        id: string,
        image?: string | null,
        name?: string | null,
        expiry_date?: string | null,
        driverID: string,
        createdAt: string,
        updatedAt: string,
        _version: number,
        _deleted?: boolean | null,
        _lastChangedAt: number,
      } | null >,
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    fcmToken?: string | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};